+++ src/record.h
@@ -3,12 +3,12 @@
 typedef struct Record {
   const char *path;
   double n_visits;
-  int last_visit;
+  long long last_visit;
 } Record;
 
 void parse_record(char *string, Record *rec);
 
-void update_record(Record *rec, int now);
+void update_record(Record *rec, long long now);
 
 char *record_to_string(Record *rec);
